Labyrinthe Updates: =============== - Added descriptions for accessory items - Added images for accessory items - Working on adding the accessory images to the view controllers - Devising a way to use accessory items efficiently.
===============

Accessory items could be used in a two ways:

1 at a time:
---------------
Accessory items are used one at a time. Purchased from the store and utilize the accessory slot on the player. Certain items are considered consumable. The consumable items are denoted by an icon that displayed whether or not the player has them. Once the consumable items are used, the icon disappears and the item no longer modifies gameplay. Non-consumable items are displayed by name in the accessory slot akin to sword/shield items. These items are permanent and are only removed when a player purchases a new item. Items are randomized by the shopkeeper. Some items may affect gameplay to the point where they need to be sorted into different blocks to combat more powerful creatures.

Collected:
-------------
Items are collected. Both consumable items and non-consumable accessory items are displayed by icons on the player's view controller. If an item is purchased and is permanent, it will be removed for the possible pool of items available from the merchant's store. The non-consumable items have a permanent effect on the player's game. The items can only be removed on death or restarting the game. Consumable items work identical to the '1 at a time approach' and are reinserted into the possible choices of purchase when they are consumed and can be purchased again to be consumed.